cleaned code
cleaned up the code. I inspected the source code from local page
next to githubs, the css was the same but I did notice .container
was waaay indented.I wouldn't think that would make a difference
but I fixed it, and other formatting issues just to be sure, and for
cleanliness.