fix link
Pointed the link on the facts page back to the landing page.
I solved my previous problem of the website not displaying
properly from git. I may be too mentally challened to code.
I had chrome ZOOMED IN to 125%, hence it displaying strangely.
At least I had narrowed down the issue to an issue with my
Chrome broweser. I thought it had something to do with the
cache not clearing properly, as it displayed properly on other
browsers, and even on Chrome on other machines. Well no
duh. Sigh...I hope somebody reads this and laughs.  In the future
I hope to be able to write better css so it won't be distorted by
only zooming in 125%.